琐碎的调整
