+++ caml-mimic-master/.gitignore
@@ -15,3 +15,5 @@ share/*
 pip-selfcheck.json
 mimicdata/mimic2/*
 mimicdata/mimic3/*
+mimicdata/backup/*
+.model_in_Section_5.4/*
