--- caml-mimic-master/learn/stacktext.py
@@ -1,5 +1,0 @@
-import torch
-y=[[torch.randn(1),torch.randn(1)],[torch.randn(1),torch.randn(1)],[torch.randn(1),torch.randn(1)]]
-print(y)
-y=torch.stack(y)
-print(y)
