+++ stacktext.py
@@ -1,0 +1,5 @@
+import torch
+y=[[torch.randn(1),torch.randn(1)],[torch.randn(1),torch.randn(1)],[torch.randn(1),torch.randn(1)]]
+print(y)
+y=torch.stack(y)
+print(y)
